fix job profile tests
