+++ src/settings/JobProfiles/tests/ViewJobProfile.test.js
@@ -20,7 +20,7 @@ import '../../../../test/jest/__mock__';
 
 import { STATUS_CODES } from '../../../utils';
 
-import { ViewJobProfile } from '../ViewJobProfile';
+import { ViewJobProfile, getAssociatedJobsURL } from '../ViewJobProfile';
 import { UploadingJobsContext } from '../../../components';
 
 const uploadContext = (canUseObjectStorage) => (
@@ -365,6 +365,19 @@ describe('ViewJobProfile component', () => {
       fireEvent.click(cancelButton[0]);
 
       await waitFor(() => expect(queryByText('Are you sure you want to run this job?')).not.toBeInTheDocument());
+    });
+  });
+
+  describe('getAssociatedJobsURL', () => {
+    const splittingTrueResources = { splitStatus: { records: [{ splitStatus: true }] } };
+    const splittingFalseResources = { splitStatus: { records: [{ splitStatus: false }] } };
+
+    it('returns splitting URL when splitting is active', () => {
+      expect(getAssociatedJobsURL(splittingTrueResources, 'splittingTrueUrl', 'splittingFalseUrl')).toEqual('splittingTrueUrl');
+    });
+
+    it('returns non-splitting URL when splitting is inactive', () => {
+      expect(getAssociatedJobsURL(splittingFalseResources, 'splittingTrueUrl', 'splittingFalseUrl')).toEqual('splittingFalseUrl');
     });
   });
 });
